refactor: move auth endpoints to separate controller/route file
