+++ packages/notes/server/controllers/users.ts
@@ -1,15 +1,10 @@
 import { Request, Response } from "express";
-import Joi, { ValidationError } from "joi";
+import { ValidationError } from "joi";
 import bcrypt from "bcrypt";
-import jwt from "jsonwebtoken";
 import { prisma } from "../lib";
 import { CustomRequest } from "../types";
+import { userSchema as schema } from "../schema";
 
-const schema = Joi.object({
-	username: Joi.string().required(),
-	password: Joi.string().required(),
-	admin: Joi.boolean(),
-});
 
 export async function getUsersNotes(req: Request, res: Response) {
 	const { username } = req.params;
@@ -19,46 +14,6 @@ export async function getUsersNotes(req: Request, res: Response) {
 	res.json(notes);
 }
 
-export async function loginOrRegister(req: Request, res: Response) {
-	try {
-		const { username, password } = await schema.validateAsync(req.body);
-		const user = await prisma.users.findOne({
-			where: { username },
-		});
-		if (user) {
-			const isValidPassword = bcrypt.compareSync(password, user.password);
-			if (isValidPassword) {
-				const token = jwt.sign({ id: user.id }, process.env.JWT_SECRET, {
-					expiresIn: "24h",
-				});
-				res
-					.cookie("token", token, { httpOnly: true })
-					.json({ ...user, password: undefined });
-			} else res.status(400).json({ message: "Invalid login credentials" });
-		} else {
-			const newUser = await prisma.users.create({
-				data: {
-					username,
-					password: bcrypt.hashSync(password, 10),
-					admin: false,
-				},
-				select: {
-					id: true,
-					username: true,
-					admin: true,
-				},
-			});
-			const token = jwt.sign({ id: newUser.id }, process.env.JWT_SECRET, {
-				expiresIn: "24h",
-			});
-			res.cookie("token", token, { httpOnly: true }).json(newUser);
-		}
-	} catch (e) {
-		if (e instanceof ValidationError) {
-			res.status(400).json({ message: e.message });
-		} else throw e;
-	}
-}
 
 export async function createUser(req: CustomRequest, res: Response) {
 	try {
