+++ packages/notes/server/routes/auth.ts
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import { loginOrRegister, logout } from "../controllers/auth";
+
+const router = Router();
+
+router.post("/login", loginOrRegister);
+router.get("/logout", logout);
+
+export default router;
