+++ packages/notes/server/routes/index.ts
@@ -1,2 +1,3 @@
+export { default as auth } from "./auth";
 export { default as notes } from "./notes";
 export { default as users } from "./users";
