+++ packages/notes/server/routes/users.ts
@@ -1,10 +1,9 @@
 import { Router } from "express";
-import { createUser, getUsersNotes, loginOrRegister } from "../controllers/users";
+import { createUser, getUsersNotes } from "../controllers/users";
 
 const router = Router();
 
 router.get("/:username/notes", getUsersNotes);
 router.post("/", createUser);
-router.post("/login", loginOrRegister);
 
 export default router;
